+++ autofit/database/migration/migration.py
@@ -1,18 +1,32 @@
+import logging
 from abc import ABC, abstractmethod
-from functools import wraps
 from hashlib import md5
+from typing import Union, Generator
 
-from sqlalchemy import text
 from sqlalchemy.exc import OperationalError
+
+from .session_wrapper import SessionWrapper
+
+logger = logging.getLogger(
+    __name__
+)
 
 
 class Identifiable(ABC):
     @property
     @abstractmethod
-    def id(self):
-        pass
+    def id(self) -> str:
+        """
+        A unique identifier generated by hashing a string
+        """
 
-    def __eq__(self, other):
+    def __eq__(
+            self,
+            other: Union["Identifiable", str]
+    ) -> bool:
+        """
+        Compares ids
+        """
         if isinstance(
                 other,
                 Identifiable
@@ -26,12 +40,74 @@ class Identifiable(ABC):
         return False
 
 
+class Step(Identifiable):
+    def __init__(self, *strings: str):
+        """
+        A collection of SQL statements executed as one step
+        in a database migration.
+
+        Parameters
+        ----------
+        strings
+            SQL statements
+        """
+        self.strings = strings
+
+    @property
+    def id(self) -> str:
+        """
+        Hash generated from underlying SQL statements
+        """
+        return md5(
+            ":".join(
+                self.strings
+            ).encode(
+                "utf-8"
+            )
+        ).hexdigest()
+
+
+class Revision(Identifiable):
+    def __init__(self, steps):
+        self.steps = steps
+
+    @property
+    def id(self):
+        return md5(
+            ":".join(
+                step.id for step
+                in self.steps
+            ).encode("utf-8")
+        ).hexdigest()
+
+    def __sub__(self, other):
+        return Revision(tuple(
+            step for step in self.steps
+            if step not in other.steps
+        ))
+
+
 class Migrator:
-    def __init__(self, *steps):
+    def __init__(
+            self,
+            *steps: Step
+    ):
+        """
+        Manages migration of an old database.
+
+        The revision table is checked to see what version a database is on.
+        This is compared to the identifier of the current revision to determine
+        the set of Steps that must be executed to migrate the database.
+
+        Parameters
+        ----------
+        steps
+            All steps recorded for every migration
+        """
         self._steps = steps
 
     @property
-    def revisions(self):
+    def revisions(self) -> Generator[Revision, None, None]:
         for i in range(1, len(self._steps) + 1):
             yield Revision(
                 self._steps[:i]
@@ -67,84 +143,3 @@ class Migrator:
                     pass
 
         wrapper.revision_id = self.latest_revision.id
-
-
-def needs_revision_table(
-        func
-):
-    @wraps(func)
-    def wrapper(self, *args, **kwargs):
-        try:
-            return func(self, *args, **kwargs)
-        except OperationalError:
-            self._init_revision_table()
-            return func(self, *args, **kwargs)
-
-    return wrapper
-
-
-class SessionWrapper:
-    def __init__(self, session):
-        self.session = session
-
-    def _init_revision_table(self):
-        self.session.execute(
-            "CREATE TABLE revision (revision_id VARCHAR PRIMARY KEY)"
-        )
-        self.session.execute(
-            "INSERT INTO revision (revision_id) VALUES (null)"
-        )
-
-    @property
-    @needs_revision_table
-    def revision_id(self):
-        for row in self.session.execute(
-                "SELECT revision_id FROM revision"
-        ):
-            return row[0]
-        return None
-
-    @revision_id.setter
-    @needs_revision_table
-    def revision_id(self, revision_id):
-        self.session.execute(
-            text(
-                f"UPDATE revision SET revision_id = :revision_id"
-            ),
-            {"revision_id": revision_id}
-        )
-
-
-class Revision(Identifiable):
-    def __init__(self, steps):
-        self.steps = steps
-
-    @property
-    def id(self):
-        return md5(
-            ":".join(
-                step.id for step
-                in self.steps
-            ).encode("utf-8")
-        ).hexdigest()
-
-    def __sub__(self, other):
-        return Revision(tuple(
-            step for step in self.steps
-            if step not in other.steps
-        ))
-
-
-class Step(Identifiable):
-    def __init__(self, *strings):
-        self.strings = strings
-
-    @property
-    def id(self):
-        return md5(
-            ":".join(
-                self.strings
-            ).encode(
-                "utf-8"
-            )
-        ).hexdigest()
